requirements.txt
openai-agents
openai
streamlit
pandas
numpy
python-dotenv
requests
